--- src/main/java/baseball/IntComparator.java
@@ -1,7 +1,0 @@
-package baseball;
-
-public class IntComparator {
-    static boolean isEqual(int num1, int num2) {
-        return num1 == num2;
-    }
-}
