+++ src/main/java/baseball/domain/ScoreCalculator.java
@@ -1,4 +1,6 @@
-package baseball;
+package baseball.domain;
+
+import baseball.utils.SystemConstant;
 
 import java.util.stream.Stream;
