+++ src/main/java/baseball/domain/ScoreManager.java
@@ -1,4 +1,8 @@
-package baseball;
+package baseball.domain;
+
+import baseball.domain.BaseballNumber;
+import baseball.domain.Score;
+import baseball.domain.ScoreCalculator;
 
 import java.util.List;
