+++ src/main/java/baseball/view/OutputHandler.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.view;
 
 import java.util.List;
