+++ src/main/java/baseball/view/OutputMsgGenerator.java
@@ -1,4 +1,6 @@
-package baseball;
+package baseball.view;
+
+import baseball.view.OutputMsgContainer;
 
 import java.util.List;
